+++ changelog.md
@@ -1,3 +1,7 @@
+# v1.0.2
+- Fixed cache setting
+- Added \[Currently WIP\] thumbnail button setting
+
 # v1.0.1
 - Added cache setting
 - Fixed thumbnails on compact lists